Fixed resize and drag and drop issues
fixes issue with drag and drop where it was not going to the cell it
was intended for. If you drag the majority of the div over the right
cell it will go right, if you drag the majority of the div over the
left cell it will go left.